Move 'authorization' token to 'customHeaders' in Jenkinsfile
Reordered the 'authorization' token location from being directly in the httpRequest to inside 'customHeaders' in the Jenkinsfile. The change was necessary as some libraries or frameworks expect the authorization token to be in the 'customHeaders'. Incorrect location could prevent the token from being recognized, consequently leading to unauthorized access issues.