Introduced new Jenkins stages for build notifications
Added two stages, 'Notificar início de build' and 'Notificar fim de build', in the Jenkinsfile to trigger notifications at the start and end of a build, given that env.RELEASE_COMMIT is not '0'. This is to improve visibility and monitoring of our build processes. Additionally, the notification text for service update has been improved for better clarity.